+++ apiserver/common/networkingcommon/spaces.go
@@ -9,9 +9,9 @@ import (
 
 	"github.com/juju/juju/apiserver/common"
 	"github.com/juju/juju/apiserver/params"
+	"github.com/juju/juju/core/network"
 	"github.com/juju/juju/environs"
 	"github.com/juju/juju/environs/context"
-	"github.com/juju/juju/network"
 )
 
 // SupportsSpaces checks if the environment implements NetworkingEnviron
